Implement geoblock for swaps on Commerce widget
